+++ CHANGELOG.md
@@ -17,9 +17,10 @@
 - Added hide Premium Flesh laying around in dungeon
 - Added Dungeon Copilot (Suggests to you what to do next in dungeon)
 - Added separate option to hide dungeon key pickup and door open messages
+- Added hide Journal Entry pages laying around in dungeon.
 
 ### Fixes
-- Fixed small bazaar Message Hider
+- Fixed message filter for small bazaar messages
 
 ## Version 0.6 - Ashfang and Summoning Mobs